chore(ci): don't update compat version on alpha releases

+++ .ci/scripts/release/compat-update.sh
@@ -1,8 +1,12 @@
-mvn versions:set-property -DgenerateBackupPoms=false -Dproperty=backwards.compat.version -DnewVersion=${RELEASE_VERSION}
-gocompat save --path=clients/go/.gocompat.json clients/go/...
+if [[ "$RELEASE_VERSION" =~ ^[0-9]+\.[0-9]+\.[0-9]+$ ]]; then
+  mvn versions:set-property -DgenerateBackupPoms=false -Dproperty=backwards.compat.version -DnewVersion=${RELEASE_VERSION}
+  gocompat save --path=clients/go/.gocompat.json clients/go/...
 
-FILE=$(mvn help:evaluate -Dexpression=ignored.changes.file -q -DforceStdout)
+  FILE=$(mvn help:evaluate -Dexpression=ignored.changes.file -q -DforceStdout)
 
-rm -f clients/java/$FILE test/$FILE exporter-api/$FILE protocol/$FILE
+  rm -f clients/java/$FILE test/$FILE exporter-api/$FILE protocol/$FILE
 
-git commit -am "chore(project): update backwards compatibility version"
+  git commit -am "chore(project): update backwards compatibility version"
+else
+  echo "Version $RELEASE_VERSION"
+fi
